Groundwork for route finding
Created a EdgeData class to better collect edge data at one point.
Added a route List to the  RoutineData where the Route from start to desination will be saved in.
Altered WaynetCrator so that it fills the EdgeData. Therefore created a List where all data is saved and a Dictionary for faster access