Refactor MonoBehaviour singletons to static classes
This allows the worlds to be loaded in the editor again, as well as make for cleaner code and scene structure. It also fixes some sneaky bugs. MeshCreator is now the default mesh creator, removing AbstractMeshCreator. WayNetManager and NpcManager are renamed to -Helper, since they don't manage anything but instead expose some helper methods.