Update ItemGrabInteractable.cs
removed alreadyGrabbedOnce and checking if rigidbody isKinematic instead. only setting the physics on if they are off right now.
assigning rigidbody on start
moved VobMeshCullingManager.I.StopTrackVobPositionUpdates(gameObject); to the bottom of the release method
sometimes there would be an error with vob mesh culling and that would stop physics from being re-enabled on item release